agents/crypto-market-agent/pyproject.toml
[tool.poetry]
name = "crypto-market-agent"
version = "0.1.0"
description = "Crypto & DeFi market intelligence agent with tiered pricing"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
strands-agents = {version = ">=1.0.0", extras = ["openai", "a2a"]}
payments-py = {version = ">=1.3.3", extras = ["strands"]}
fastapi = "^0.120.0"
uvicorn = ">=0.34.2,<1.0.0"
httpx = "^0.28.0"
openai = "^1.40.0"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
agent = "src.agent:main"
agent-a2a = "src.agent_a2a:main"
client = "src.client:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
